Development

+++ project3/TrelloV2-Angular/src/app/scrum-board-view/scrum-board-view.component.ts
@@ -197,15 +197,21 @@ export class ScrumBoardViewComponent implements OnInit {
     )
   
   }
-  removeLane(laneToRemove){
-    this.laneDislayService.deleteLane(laneToRemove).subscribe(
-      data => console.log(this.responseStatus = data),
-      err => console.log(err),
-      () => console.log('delete Lane request completed')
-    );
-    setTimeout(function () {
-      this.displayLanes();
-    }.bind(this), 1000);
+  removeLane(laneToRemove, lId : number, laneTitle: string){
+    if(confirm("Are you sure to delete \" "+ laneTitle +"\" ?" 
+              +"\nThe cards will also be deleted at the same time!")) {
+
+      this.laneDislayService.deleteLane(laneToRemove).subscribe(
+        data => console.log(this.responseStatus = data),
+        err => console.log(err),
+        () => console.log('delete Lane request completed')
+      );
+      setTimeout(function () {
+        this.displayLanes();
+      }.bind(this), 1000);
+    }
+
+   
   }
